story for subspace identification algorithm design

+++ AnalysisKernel/include/FastFourierTrans.h
@@ -30,43 +30,43 @@ typedef struct complex
 ///////////////////////////////////////////////////////////////////////////////////////////
 //function  declaration
 
-int transform(double real[], double imag[], size_t n);
+int transform(double real[], double imag[], Uint32 n);
 
-int inverse_transform(double real[], double imag[], size_t n);
+int inverse_transform(double real[], double imag[], Uint32 n);
 
-int transform_radix2(double real[], double imag[], size_t n);
+int transform_radix2(double real[], double imag[], Uint32 n);
 
-int transform_bluestein(double real[], double imag[], size_t n);
+int transform_bluestein(double real[], double imag[], Uint32 n);
 
-int convolve_real(double x[], double y[], double out[], size_t n);
+int convolve_real(double x[], double y[], double out[], Uint32 n);
 
-int convolve_complex(double xreal[], double ximag[], double yreal[], double yimag[], double outreal[], double outimag[], size_t n);
+int convolve_complex(double xreal[], double ximag[], double yreal[], double yimag[], double outreal[], double outimag[], Uint32 n);
 
 double	IFatanCal(double xq, double xd);
 
 // Private function prototypes
 static int32	reverse_bits(int32 x, Uint32 n);
-static void		*memdup(void *src, size_t n);
+static void		*memdup(void *src, Uint32 n);
 static int32	reverse_bits1(int32 x, Uint32 levels);
 
 ///////////////////////////////////////////////////////////////////////////////////////////
 //new function 
 
-int transform_x(COMPLEX* x, size_t n);
+int transform_x(COMPLEX* x, Uint32 n);
 
-int inverse_transform_x(COMPLEX* x, size_t n);
+int inverse_transform_x(COMPLEX* x, Uint32 n);
 
-int transform_radix2_x(COMPLEX* x, size_t n);
+int transform_radix2_x(COMPLEX* x, Uint32 n);
 
-int transform_bluestein_x(COMPLEX* x, size_t n);
+int transform_bluestein_x(COMPLEX* x, Uint32 n);
 
-int transform2D(COMPLEX* X1, COMPLEX* X2, size_t N);
+int transform2D(COMPLEX* X1, COMPLEX* X2, Uint32 N);
 
-int inverse_transform2D(COMPLEX* X1, COMPLEX* X2, size_t N);
+int inverse_transform2D(COMPLEX* X1, COMPLEX* X2, Uint32 N);
 
-int transform1(COMPLEX* X, size_t N);
+int transform1(COMPLEX* X, Uint32 N);
 
-int inverse_transform1(COMPLEX* X, size_t N);
+int inverse_transform1(COMPLEX* X, Uint32 N);
